edit deletion messages

+++ server/controllers/fileRouter.js
@@ -106,10 +106,16 @@ fileRouter.post('/deleteUser/'), async (req,res) => {
   const decodedToken = auth.decodeToken(token)
   const id = decodedToken.id
 
+  if(id !== null){
+
   await Company.find({'user': `${id}`}).deleteMany()
   await Employee.find({'user': `${id}`}).deleteMany()
   await User.findByIdAndDelete(id)
-  
+
+  res.send(`User ${id} deleted`)
+  }else{
+    res.send('Error in user deletion')
+  }
 }
 
 module.exports = fileRouter
